sidekiq: disable update checker (#1368)
This is configured to look for upstream Mastodon updates, so it isn't
actually doing anything useful for Hometown admins. Eventually it can be
turned back on if there's a Hometown-specific API that returns the
appropriate versions, but in the meantime this feature is misleading.

Looking at the sidekiq-scheduler docs, I think this should be enough to
turn it off:
https://github.com/sidekiq-scheduler/sidekiq-scheduler?tab=readme-ov-file#schedule-configuration

I also disabled it in the place where upstream Mastodon configures the
feature to be on or off:
https://docs.joinmastodon.org/admin/upgrading/#automated_checks

Fixes #1367.